EditDialog: refactor - extract types
