+++ src/components/FeaturePanel/Climbing/useSaveCragFactory.tsx
@@ -8,10 +8,10 @@ import {
   getWikimediaCommonsKey,
 } from './utils/photo';
 import { Setter } from '../../../types';
-import { DataItem } from '../EditDialog/context/useEditItems';
 import { saveChanges } from '../../../services/osm/auth/osmApiAuth';
 import { fetchFreshItem } from '../EditDialog/context/itemsHelpers';
 import { stringifyPath } from './utils/pathUtils';
+import { DataItem } from '../EditDialog/context/types';
 
 const getUpdatedPhotoTags = (route: ClimbingRoute) => {
   const updatedTags = {};
