+++ src/components/FeaturePanel/EditDialog/EditContent/FeatureEditSection/LocationEditor/useStaticMarkers.tsx
@@ -5,9 +5,9 @@ import { Button, Stack, Typography } from '@mui/material';
 import { createMapEffectHook } from '../../../../../helpers';
 import { t } from '../../../../../../services/intl';
 import { isGpsValid } from './isGpsValid';
-import { EditDataItem } from '../../../context/useEditItems';
 import { useEditContext } from '../../../context/EditContext';
 import { Setter } from '../../../../../../types';
+import { EditDataItem } from '../../../context/types';
 
 const GRAY_MARKER = {
   color: '#555',
