+++ src/components/FeaturePanel/EditDialog/context/__tests__/useEditItems.convertToRelationFactory.test.tsx
@@ -3,7 +3,8 @@ import { fetchWays } from '../../../../../services/osm/fetchWays';
 import { addEmptyOriginalState, fetchFreshItem } from '../itemsHelpers';
 import { getNewId } from '../../../../../services/getCoordsFeature';
 
-import { convertToRelationFactory, DataItem } from '../useEditItems';
+import { convertToRelationFactory } from '../useEditItems';
+import { DataItem } from '../types';
 
 jest.mock('../../../../../services/osm/fetchParentFeatures', () => ({
   fetchParentFeatures: jest.fn(),
