+++ src/components/FeaturePanel/EditDialog/context/useEditItems.tsx
@@ -1,4 +1,4 @@
-import { FeatureTags, LonLat } from '../../../../services/types';
+import { LonLat } from '../../../../services/types';
 import { getApiId } from '../../../../services/helpers';
 import { Setter } from '../../../../types';
 import { useCallback, useMemo, useState } from 'react';
@@ -10,49 +10,16 @@ import { fetchParentFeatures } from '../../../../services/osm/fetchParentFeature
 import { fetchWays } from '../../../../services/osm/fetchWays';
 import { addEmptyOriginalState, fetchFreshItem } from './itemsHelpers';
 import { isEqual } from 'lodash';
-
-export type TagsEntries = [string, string][];
-
-export type Members = Array<{
-  shortId: string;
-  role: string;
-  label: string; // cached from other dataItems, or from originalFeature
-  // TODO rename to originalLabel - only to be used when member is not among editItems
-}>;
-
-// internal type stored in the state
-export type DataItem = {
-  shortId: string;
-  version: number | undefined; // undefined for new item
-  tagsEntries: TagsEntries;
-  toBeDeleted: boolean;
-  nodeLonLat: LonLat | undefined; // only for nodes & for relation converted from node
-  nodes: number[] | undefined; // only for ways
-  members: Members | undefined; // only for relations
-  originalState: {
-    tags: FeatureTags;
-    isDeleted: boolean;
-    nodeLonLat: LonLat | undefined;
-    nodes: number[] | undefined;
-    members: Members;
-  };
-};
-
-export type DataItemRaw = Omit<DataItem, 'originalState'>;
-
-export type EditDataItem = DataItem & {
-  setTagsEntries: SetTagsEntries;
-  tags: FeatureTags;
-  setTag: (k: string, v: string) => void;
-  presetKey: string;
-  presetLabel: string;
-  setMembers: SetMembers;
-  setShortId: SetShortId;
-  setNodeLonLat: (lonLat: LonLat) => void;
-  toggleToBeDeleted: () => void;
-  convertToRelation: ConvertToRelation;
-  modified: boolean;
-};
+import {
+  ConvertToRelation,
+  DataItem,
+  EditDataItem,
+  Members,
+  SetMembers,
+  SetShortId,
+  SetTagsEntries,
+  TagsEntries,
+} from './types';
 
 export const isInItems = (items: DataItem[], shortId: string) =>
   items.some((item) => item.shortId === shortId);
@@ -178,7 +145,6 @@ const fetchParentItems = async (shortId: string) => {
   );
 };
 
-type ConvertToRelation = () => Promise<string>;
 const convertToRelationFactory = (
   setData: Setter<DataItem[]>,
   shortId: string,
@@ -239,7 +205,6 @@ const convertToRelationFactory = (
   };
 };
 
-type SetTagsEntries = (updateFn: (prev: TagsEntries) => TagsEntries) => void;
 const setTagsEntriesFactory =
   (setDataItem: SetDataItem, tagsEntries: TagsEntries): SetTagsEntries =>
   (updateFn) =>
@@ -248,7 +213,6 @@ const setTagsEntriesFactory =
       tagsEntries: updateFn(tagsEntries),
     }));
 
-type SetShortId = (shortId: string) => void;
 const setShortIdFactory =
   (setDataItem: SetDataItem): SetShortId =>
   (shortId) =>
@@ -257,7 +221,6 @@ const setShortIdFactory =
       shortId: shortId,
     }));
 
-type SetMembers = (updateFn: (prev: Members) => Members) => void;
 const setMembersFactory =
   (setDataItem: SetDataItem, members: Members): SetMembers =>
   (updateFn) =>
